+++ rust/substreams/src/lib.rs
@@ -1,4 +1,6 @@
 mod externs;
+mod hex;
+
 pub mod log;
 pub mod memory;
 pub mod pb;
@@ -6,9 +8,9 @@ pub mod proto;
 pub mod rpc;
 pub mod state;
 
-pub fn output<M: prost::Message>(msg: M) {
-    log::debug!("value");
+pub use crate::hex::Hex;
 
+pub fn output<M: prost::Message>(msg: M) {
     // Need to return the buffer and forget about it issue occured when trying to write large data
     // wasm was "dropping" the data before we could write to it, which causes us to have garbage
     // value. By forgetting the data we can properly call external output function to write the