+++ wasm/module.go
@@ -439,33 +439,6 @@ func (m *Module) registerStateImports(imports *wasmer.ImportObject, store *wasme
 		},
 	)
 
-	functions["sum_bigfloat"] = wasmer.NewFunction(
-		store,
-		wasmer.NewFunctionType(
-			Params(wasmer.I64 /* ordinal */, wasmer.I32, wasmer.I32 /* key */, wasmer.I32, wasmer.I32 /* value */),
-			Returns(),
-		),
-		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			if m.CurrentInstance.outputStore == nil && m.CurrentInstance.updatePolicy != pbsubstreams.Module_KindStore_UPDATE_POLICY_SUM && m.CurrentInstance.valueType != "bigfloat" {
-				return nil, fmt.Errorf("invalid store operation: 'sum_bigfloat' only valid for stores with updatePolicy == 'sum' and valueType == 'bigfloat'")
-			}
-			ord := args[0].I64()
-			key, err := m.CurrentInstance.heap.ReadString(args[1].I32(), args[2].I32())
-			if err != nil {
-				return nil, fmt.Errorf("reading string: %w", err)
-			}
-			value, err := m.CurrentInstance.heap.ReadString(args[3].I32(), args[4].I32())
-			if err != nil {
-				return nil, fmt.Errorf("reading bytes: %w", err)
-			}
-
-			toAdd, _, err := big.ParseFloat(value, 10, 100, big.ToNearestEven)
-			m.CurrentInstance.outputStore.SumBigFloat(uint64(ord), key, toAdd)
-
-			return nil, nil
-		},
-	)
-
 	functions["set_min_int64"] = wasmer.NewFunction(
 		store,
 		wasmer.NewFunctionType(